+++ CHANGELOG.md
@@ -1,7 +1,10 @@
-## X.X.X
+## x.x.x x-x-x
 
 ### PaymentSheet
 * [Fixed] The Apple Pay logo in `PaymentOptionDisplayData` no longer contains additional padding, bringing it in line with other payment method logos.
+
+### EmbeddedPaymentElement
+* [Changed] When using `EmbeddedPaymentElement.Configuration.FormSheetAction.confirm`, the completion block is now called with a `canceled` result if the user closes the form sheet without completing the transaction.
 
 ## 24.19.0 2025-08-04